Move fixture_dir helper
