+++ rust/geoarrow-geoparquet/src/test/geoarrow_data/mod.rs
@@ -1,0 +1,1 @@
+mod example;
